requirements.txt
streamlit
pytesseract
pillow
opencv-python
numpy
pandas
openpyxl
opencv-python-headless
pdf2image
